backend/requirements.txt
fastapi
uvicorn
gitpython
python-multipart
requests
openai 